refactor(MyInfo): 로그아웃 모달 추가 및 컴포넌트 분리
