+++ src/pages/MyInfoEditPage/index.jsx
@@ -51,7 +51,7 @@ const MyInfoEditPage = () => {
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    if (!passwordInvalid && !confirmInvalid) {
+    if (password && confirm && !passwordInvalid && !confirmInvalid) {
       onChangePassword(password);
       setIsModal(true);
     }
@@ -66,8 +66,8 @@ const MyInfoEditPage = () => {
               marginLeft: 20,
               textAlign: 'left',
               fontWeight: 700,
-              fontSize: 26,
             }}
+            fontSize={26}
             block={true}
           >
             비밀번호를 설정해주세요
